+++ dashboard/managers/graphs.py
@@ -146,6 +146,7 @@ class GraphManager(BaseManager):
     def _format_stats_for_default_graphs(self, locale_sequence, stats_dict, desc):
         """
         Formats stats dict for graph-ready material
+        - sorting and normalization
         """
         stats_for_graphs_dict = OrderedDict()
         stats_for_graphs_dict['pkg_desc'] = desc
@@ -178,6 +179,9 @@ class GraphManager(BaseManager):
         if not package:
             return {}
         lang_id_name, stats_dict, pkg_desc = self.package_manager.get_trans_stats(package)
+        upstream_stats = self.package_manager.get_upstream_stats(package)
+        if upstream_stats:
+            stats_dict['Upstream'] = upstream_stats
         # format trans_stats_list for graphs
         return self._format_stats_for_default_graphs(lang_id_name, stats_dict, pkg_desc)