+++ dashboard/tests/test_inventory.py
@@ -1,0 +1,130 @@
+# Copyright 2017 Red Hat, Inc.
+# All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License"); you may
+# not use this file except in compliance with the License. You may obtain
+# a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations
+# under the License.
+
+from django.test import TestCase
+from dashboard.managers.inventory import InventoryManager
+from dashboard.models import Languages, LanguageSet, TransPlatform
+
+
+class InventoryTest(TestCase):
+
+    inventory_manager = InventoryManager()
+
+    def setUp(self):
+        """
+        Lets dump some test data in db
+        """
+        Languages.objects.update_or_create(
+            locale_id='ja_JP', lang_name='Japanese', locale_alias='ja', lang_status=True
+        )
+
+        Languages.objects.update_or_create(
+            locale_id='fr_FR', lang_name='French', locale_alias='fr', lang_status=False
+        )
+
+        LanguageSet.objects.update_or_create(
+            lang_set_name='Custom Set', lang_set_slug='custom-set', lang_set_color='Peru',
+            locale_ids=['fr_FR', 'ja_JP']
+        )
+
+        TransPlatform.objects.update_or_create(
+            engine_name='zanata', subject='public', api_url='https://translate.zanata.org',
+            platform_slug='ZNTAPUB', server_status=True
+        )
+
+    def test_get_locales(self):
+        """
+        Test get_locales
+        """
+        japanese_locale = self.inventory_manager.get_locales(pick_locales=['ja_JP'])
+        self.assertEqual(japanese_locale[0].lang_name, 'Japanese')
+        self.assertEqual(japanese_locale[0].locale_alias, 'ja')
+
+    def test_get_locale_alias(self):
+        """
+        Test get_locale_alias
+        """
+        locale_alias = self.inventory_manager.get_locale_alias('fr_FR')
+        self.assertEqual(locale_alias, 'fr')
+
+        locale_alias = self.inventory_manager.get_locale_alias('de_DE')
+        self.assertEqual(locale_alias, 'de_DE')
+
+    def test_get_locales_set(self):
+        """
+        Test get_locales_set
+        """
+        active_locales, inactive_locales, aliases = \
+            self.inventory_manager.get_locales_set()
+        self.assertEqual(len(active_locales), 1)
+        self.assertEqual(len(inactive_locales), 1)
+        self.assertEqual(len(aliases), 2)
+
+    def test_get_langset(self):
+        """
+        Test get_get_langset
+        """
+        lang_set = self.inventory_manager.get_langset(langset_slug='custom-set')
+        self.assertEqual(lang_set.lang_set_name, 'Custom Set')
+        self.assertEqual(lang_set.lang_set_color, 'Peru')
+
+    def test_get_langsets(self):
+        """
+        Test get_langsets
+        """
+        lang_sets = self.inventory_manager.get_langsets(
+            fields=['lang_set_name', 'locale_ids']
+        )
+        self.assertEqual(len(lang_sets), 1)
+        self.assertNotIn('lang_set_color', vars(lang_sets[0]))
+        self.assertListEqual(lang_sets[0].locale_ids, ['fr_FR', 'ja_JP'])
+
+    def test_get_locale_groups(self):
+        """
+        Test get_locale_groups
+        """
+        locale_groups = self.inventory_manager.get_locale_groups('ja_JP')
+        self.assertDictEqual(locale_groups, {'ja_JP': ['custom-set']})
+
+    def test_get_all_locales_groups(self):
+        """
+        Test get_all_locales_groups
+        """
+        groups_of_all_locales = self.inventory_manager.get_all_locales_groups()
+        self.assertDictEqual(groups_of_all_locales,
+                             {'ja_JP': ['custom-set'], 'fr_FR': ['custom-set']})
+
+    def test_get_translation_platforms(self):
+        """
+        Test get_translation_platforms
+        """
+        transplatforms = self.inventory_manager.get_translation_platforms(engine='zanata')
+        self.assertEqual(transplatforms[0].api_url, 'https://translate.zanata.org')
+        self.assertEqual(transplatforms[0].platform_slug, 'ZNTAPUB')
+
+    def test_get_transplatforms_set(self):
+        """
+        Test get_transplatforms_set
+        """
+        active_platforms, inactive_platforms = self.inventory_manager.get_transplatforms_set()
+        self.assertEqual(len(active_platforms), 1)
+        self.assertEqual(len(inactive_platforms), 0)
+
+    def test_get_transplatform_slug_url(self):
+        """
+        test get_transplatform_slug_url
+        """
+        slug_url_tuple = self.inventory_manager.get_transplatform_slug_url()
+        self.assertTupleEqual(slug_url_tuple, (('ZNTAPUB', 'https://translate.zanata.org'),))